trying to fix some C/X stuff

+++ tools/create_amiga_archive.py
@@ -6,59 +6,67 @@ progdir = pathlib.Path(__file__).parent.parent.absolute()
 # gadf from https://github.com/sphair/gadf
 
 gamename = "vulgus"
+
+create_dist = True
+create_floppy = True
+
 # JOTD path for cranker, adapt to whatever your path is :)
 os.environ["PATH"] += os.pathsep+r"K:\progs\cli"
 
-for s in ["convert_sounds.py","convert_graphics.py"]:
-    subprocess.check_call(["cmd","/c",s],cwd=os.path.join(progdir,"assets","amiga"))
+if create_dist:
+    for s in ["convert_sounds.py","convert_graphics.py"]:
+        subprocess.check_call(["cmd","/c",s],cwd=os.path.join(progdir,"assets","amiga"))
 
-cmd_prefix = ["make","-f",os.path.join(progdir,"makefile.am")]
+    cmd_prefix = ["make","-f",os.path.join(progdir,"makefile.am")]
 
-subprocess.check_call(cmd_prefix+["clean"],cwd=progdir /"src")
+    subprocess.check_call(cmd_prefix+["clean"],cwd=progdir /"src")
 
-subprocess.check_call(cmd_prefix+["RELEASE_BUILD=1"],cwd=progdir /"src")
-# create archive
+    subprocess.check_call(cmd_prefix+["RELEASE_BUILD=1"],cwd=progdir /"src")
+    # create archive
 
-outdir = progdir / "dist" / f"{gamename}_HD"
+    outdir = progdir / "dist" / f"{gamename}_HD"
 
-if os.path.isdir(outdir):
-    shutil.rmtree(outdir)
+    if os.path.isdir(outdir):
+        shutil.rmtree(outdir)
 
-outdir.mkdir(exist_ok=True,parents=True)
+    outdir.mkdir(exist_ok=True,parents=True)
 
-for file in ["readme.md",f"{gamename}.slave"]:
-    shutil.copy(progdir / file,outdir)
+    for file in ["readme.md",f"{gamename}.slave"]:
+        shutil.copy(progdir / file,outdir)
 
-assets = progdir /"assets"/"amiga"
-shutil.copy(assets/f"{gamename.title()}.info",outdir)
-
-
-
+    assets = progdir /"assets"/"amiga"
+    shutil.copy(assets/f"{gamename.title()}.info",outdir)
 
 
-for ext in [""]:
-    exename = f"{gamename}{ext}"
-    shutil.copy(progdir/exename,outdir)
-    # ATM both packers generate incorrect exe. propack seems to do better
-    #subprocess.run(["cranker_windows.exe","-f",progdir/exename,"-o",progdir/f"{exename}.rnc"],check=True)
-    #subprocess.run(["shrinkler","-p","-c",progdir/exename,progdir/f"{exename}.rnc"],check=True)
-subprocess.run(cmd_prefix+["clean"],cwd=progdir/"src",check=True)
 
-arcname = progdir / f"{gamename}_HD.lha"
-arcname.unlink(missing_ok=True)
-cmd = ["lha","-r","a",arcname,"*"]
 
-subprocess.run(cmd,cwd=outdir.parent,check=True)
 
-# create floppy
-if False:
     for ext in [""]:
         exename = f"{gamename}{ext}"
-        shutil.move(progdir/f"{exename}.rnc",progdir/exename)
+        shutil.copy(progdir/exename,outdir)
+        # ATM both packers generate incorrect exe. propack seems to do better
+        #subprocess.run(["cranker_windows.exe","-f",progdir/exename,"-o",progdir/f"{exename}.rnc"],check=True)
+        #subprocess.run(["shrinkler","-p","-c",progdir/exename,progdir/f"{exename}.rnc"],check=True)
+    subprocess.run(cmd_prefix+["clean"],cwd=progdir/"src",check=True)
+
+    arcname = progdir / f"{gamename}_HD.lha"
+    arcname.unlink(missing_ok=True)
+    cmd = ["lha","-r","a",arcname,"*"]
+
+    subprocess.run(cmd,cwd=outdir.parent,check=True)
+
+# create floppy
+if create_floppy:
+    for ext in [""]:
+        exename = f"{gamename}{ext}"
+        # real rnc file done using propack on amiga. Yes, it sucks but it's the only way to create a floppy version
+        shutil.copy(progdir/f"{exename}.rnc",progdir/exename)
 
     #shutil.copy(assets/"disk.info",progdir)
     adf_name = gamename.title()+".adf"
-    cmd = ["gadf","-i",gamename,"-a",adf_name,"-l",gamename.title(),"readme.md"]
+    with (progdir/"floppy").open("w"):
+        pass
+    cmd = ["gadf","-i",gamename,"-a",adf_name,"-l",gamename.title(),"readme.md","floppy"]
     subprocess.run(cmd,cwd=progdir,check=True)
 
     # create a .zip for the floppy
